crds: fix check for updated CRDs on upgrades
There is a small delay between CRDs being created/updated, and the ApiServer
being ready to serve that API version. The existing check only works for the
initial creation of the API, but not if upgrading CRDs with new version.

The fix is to directly query for the API Resource in a specific version. This
requires upgrading the kubernetes client, as the required client function is
not available in 5.8.0.