crd: set request retries for client
Sometimes the Kubernetes API is not available, or the connection cannot
be established for whatever reason. Since the API is usually available again
after a short time, many clients just retry the command after a little while.

The java client we use also supports this, we just have to enable it. As a
start, use a maximum of 3 retries before failing the request.