docker: fix lvm config manipulation
Manipulation of lvm.conf failed after RHEL decided to update their lvm.conf:
instead of explicitly setting all values, they remain unset, with only a hint
of the default value in a comment. Our manipulation then just changed this
default value comment, leaving the lvm behviour unchanged.

Instead of relying on sed manipulation, use lvmconfig to merge the current
configuration with the changes we want to make. This should make our change
more robust against changes in the original lvm.conf.

As a side-effect, this also strips all the existing comments out of the file,
but that should not really matter in a container, as the file is not persistent
in any case.