Sos: First iteration of rework
This iteration fixes the issue with 16MB limitation, but is still
not finished. The satellite will send all of the SOS report to the
controller. If the controller cannot write the data faster to
disk than it receives it from the satellite, the controller's
TCP thread will run into an OutOfMemoryError.

The 16MB issue was simply solved by splitting every file from the
SOS report that is larger than 10MB into 10MB chunks.