crds: wait for CR api using actual API
Testing on exoscale revealed that the exoscale API server reports the
discovery API being available before the actual API is ready. The best
guess as to why this happens is that there may be some proxy-ing going
on.

This caused the "CRD ready" code to report ready before the API could
be used.

This is similar to 5d60927d55b8 where we already tried to fix that issue
once.