Auth: Start pinging earlier
There are rare cases where the satellite is still busy with
authentication and/or applying FullSync where the controller
is still not pinging the satellite. If in this time-frame the
connection breaks, the controller will not add the satellite to
the ReconnectorTask (as usually that is done by the PingTask,
but in the current time-frame the satellite is not yet in the
PingTask). As a result, the satellite stays offline and the
controller will never try to reestablish a new connection.

This problem can also be manually solved with a simple
'linstor node reconnect <nodename>'.