ScopeRunner: Do not release locks before rollback
This patch attempts to fix very rare ConcurrentMoficiationExceptions.
The scenario is that a node is offline but we still try to create
a snapshot. For some reasons we do not fail fast enough, only at
the step where we send out the updates to the satellites. That
"PeerNotConnectedException" basically makes the first flux-step
fail (expected). The scopeRunner leaves the try block, executes the
finally block, which starts with giving up the locks. After the
locks were given up the transaction is rolled back. In our scenario
this removes the SnapshotDefinitions from the corresponding
ResourceDefinition.

If concurrently the client queries at the "right time" a
'linstor snapshot list', these two thread cause a
ConcurrentModificationException.