GenSql: Load and inject JclCrypto to migrations
Some migrations requite a cryptoProvider which caused NPEs so far
as the GenerateSql does not load things like the Controller (no
guice).

The current solution relies on the fact that the jclcrypto project
is a sibling directory of the controller project (where the
GenreateSql class is located in) in order to find the corresponding
compiled JCLCryptoProvider.class.