docker: manually start lvm daemon
dmeventd will be started by lvm commands (i.e. lvcreate, lvremove, ..)
if it does not already run. This is an issue if started in a context with
a high soft file limit: dmeventd will try to close _all_ FDs after forking.
The limit on RHEL9 is higher than 1_000_000_000, which caused the initial
command to wait for over 100 seconds, mostly leading to linstor commands
timing out.

The workaround here is to start "dmeventd" before starting the satellite
with a known reasonable file limit.