Sos: second iteration
The problem with the first iteration was, that if the satellite
sent the data faster than the controller can save it to disk, the
controller's TCP thread runs out of memory and dies. This kill
all satellite communication. As the rest of the controller might
still be up and running, the satellites might even auto-evict and
worse things might happen.

To solve this issue the controller now first requests a list of
files (name + size, no content). While the satellite assembles this
list, it also copies the data into a satellite-local folder in
/var/lib/linstor.d/sos/... so that we have a copy instead of having
to worry about that the controller might request missing files.

The controller is also able to request specific chunks of files
(offset + length based) and limits the requests to ~10MB. This
should still be enough with all the other overhead before reaching
the 16MB hard limit of linstor.