Sos: Do not copy ErrorReports
ErrorReports usually do not move and are not temporary (like the
output of commands as 'dmesg', journal). Therefore we do not need to make
a copy of all ErrorReports for each SosReport, but we can simply send
the controller reading from the original ErrorReport.

This change also requires to send the controller absolute file paths
instead of relative ones.