Resize: Attempt to fix stuck resize
The issue this commit addresses is that if a resource got stuck
in resizing and the controller gets restarted, the nextFlux is
lost. Losing the nextFlux causes the "finishResize" method to
never be called, even if the satellite actually managed to resize
the resource.
As the "finishResize" method removes the RESIZE-flags, not calling
that method leads to the issue.

The solution was already half-way implemented, namely using the
CtrlSatelliteConnectionListener, which gets called when a
resource/-definition gets connected. In this event, all VlmDfns
are checked for the RESIZE flag and the updateSatellites method
is called to continue. However, that method also did not call
the finishResize in case of a DRBD resource.

This patch unifies the good case with the bad case (broken resize,
ctrl restart, etc.. => CtrlSatelliteConnectionListener-path) so that
in both cases the same logic is executed, including the final
finishResize.

Additionally, the different "DRBD before others" in case of shrinking
was only considered in the good case, not in the bad case. This should
also be fixed with the unification and using of a new RESIZE_SHRINK
flag on VlmDfn.