Layer: Add IgnoreReason
In some cases a layer should be entirely skipped on the satellite
while processing. For example even if layer-stack states there should
be a DRBD above NVME, the satellite that has the NVME target resource
deployed does not necessarily need to have DRBD installed at all.
However the DrbdLayer class starts its processing with trying to
sync Linstor's DRBD state with the actual DRBD state (which fails
if DRBD is not installed).

This patch adds an "ignoreReason" (null / empty means no reason,
i.e. the layer should not be skipped), where the DrbdRscData gets
marked by the NVMe layer (on controller side, during rscData-creation)

Setting a resource INACTIVE requires now an additional step where
the INACTIVE flag as well as the INACTIVATING flags are set.
That is needed so that the layers are *not* ignored while INACTIVATING
so that they can deconstruct themself before getting skipped