autoplace: only migrate tiebreaker to eligible diskless
Previously, as soon as a diskless was created, the tiebreaker was removed,
i.e. "migrated" to the diskless. This caused some loss of redundancy in
situations where constraints via "replicas-on-different" was used. A diskless
did not need to fullfil the normal autoplace criteria, so the tie breaker might
have been moved to the same zone as a full replica, decreasing the failure
tolerance.

The fix implemented here is to decide if a diskless counts as adequat
replacement for the tiebreaker by checking if "the autoplacer could have
chosen that node".