BackupShipping: fix backup schedule delete
When calling 'backup schedule delete' while a shipping
is in progress, the delete was unable to stop following
scheduled shippings from starting because as soon as the
ongoing shipment finished it re-added the shipping config
to the list of active shippings, essentially undoing what
the delete-command did.
One way to notice this problem (aside from noticing that
the shippings do not stop) is to check 'schedule lbr'.
There will be two entries per rsc that has this problem,
one noting the active shipping, and one with the reason
as 'none set', showing that any scheduled-shipping-props
are now gone.