DB: Improve DB-update-logging
These changes make it possible to specify two lambdas for logging:
One (that has already existed for a while) uses the object in
question from the parent, while the other (the new one) uses the
data to be updated. This is needed so that complex objects that do not
have a toString implementation that can be easily compared to the
logging-string the first method creates can still provide useful
logs.
For all other objects, the toString method is the default for this
new parameter.