BackupShipping: abort failed shippings after ctrl-restart
Make sure to remove the SHIPPING flag and special props related
to non-persistent data (specifically to stlt-remotes) from the snaps
after the ctrl is restarted.
Also sets the SHIPPING_ABORT flag to mark the failure more clearly.
This is a cleanup that only the ctrl needs to do since the stlt(s)
already abort any shippings as soon as they lose connection with
the ctrl.