BackupShipping: return copy instead of actual map
Make sure the BackupInfoManager only returns copies of maps instead
of the actual map itself. This will prevent accidental changes to
the actual data as well as solve a ConcurrentModificationException
that used to occur in
CtrlSnapshotShippingAbortHandler.abortBackupShippings().
Also changes an interation in the AbsBackupShippingService to be over
a copy instead of the actual map to make sure the .clear() does not
cause an exception in other parallel runs of this method.