BackupShipping: Allow shipping of rsc to several remotes at once
Shipping the same rsc to several remotes at once is now possible
(even though it should have been a while ago). This also
introduced the need to only abort shippings to a single remote
instead of all shippings of the given rsc.
Some fixes went along with that change, most importantly the
destination in an l2l-shipping also needs to save the name of the
l2l-remote in the stlt-remote for aborts to work.