CrtVlm: Fix size-check for storage pools
This is more a workaround than an actual fix.
The problem is also described as a java comment, but basically the
snapshot (and backup) restore have information about the required
meta-sizes for the meta-storagepools, whereas the usual resource
create does not have this information since it is only calculated
on the satellite.

This workaround skips the size-check for meta-pools instead of
correctly checking also for the resource create case. A proper fix
would require to move (at least a part of) the size calculation
to the controller