physical-storage: set VG during thin pool creation
The pool name given on device pool creation worked differently than
the pool name used when creating a storage pool, at least for thin LVM.

This commit fixes this decrepancy by checking if the given pool name already
contains a VG prefix. If it is present it is used, if not we fall back to
the generated "linstor_$pool" VG name.