Storage: do not wipefs on initiator
Since an initiator only connects to an existing target, it does not
make sense to wipefs the data.

The current use-case was restoring an EBS snapshot into a new
resource with DRBD data. The EBS initiator did not even knew that
this was (more or less) part of a snapshot restore operation and
just wipefs-ed the DRBD magic bytes, causing the DrbdLayer to
recreate metadata which caused the DRBD resource to never get
upToDate.