crd: implement local caching of resources
Since we can't efficiently index k8s resources, some operations required
the loading of all k8s resources of a specific type. Notabely, this included
the PropsContainers, which mean that having lots of properties would slow down
many operations. In extreme cases, this meant waiting for >30m when initially
loading the database.

To mitigate this issue, implement a simple caching client. Since a single
LINSTOR Controller should be the only one making any kind of modification
on the DB resources, we can simply use a local ConcurrentHashMap to ensure
the information is in sync with the cluster.

While the issue is limited to those resource we continously need to list
in order to find a specific subset of resources, it is simpler to
cache all resources instead.