EBS: Fix make-available support
It almost worked already but the check if DRBD is available was
broken (not just for EBS) since Linstor only checked for "is
DRBD contained in the layerstack". In case of a single EBS-target
this might be true, but creating a diskless DRBD still does not
work, since the EBS target resource does not have a DRBD on top.

This patch changed the check for empty ignore reason for the DRBD
resource data.