AllUpToDateCheck: Add null check
A list of nodes with a resource name might be checked if all
resources are UpToDate. However creating a snapshot checks if
all resources from the exisiting snapshots are UpToDate. That could
mean that there might be snapshots which no longer have a
corresponding resource (since a resource can be deleted after the
snapshot was taken). Hence the null check