EBS: Prevent EBS modifications during cooldown
An EBS volume has by default 6 hours of cooldown period. During
that time resize and other volume-modification requests are
rejected by Amazon.

To prevent to let Linstor being stuck in a resize for the entire
cooldown period, we prevent resize and a few other EBS actions
already on the controller