gradle: upgrade okhttp to fix IPv6 TLS certs
okhttp, as used by our k8s client, had a bug that caused certificate
validation to fail when connecting via IPv6 address. The bug was caused
by not using the canonical form of the IP address.

To fix the issue, we instruct gradle to select a newer version of okhttp,
which contains a fix for the issue. This new version is not the default
dependency due to some concerns from the fabrics.io team regarding
the use of kotlin for newer okhttp versions.