NodeCon: Changed Proto serialization
NodeConnPojo included a bunch of otherNode* field for the case that the
satellites receives a node connection to an unknown node and needs to
create that new node.

That was simplified with simply serializing a large part of the other node.
The only exception is that the other node's node-connections are not
serialized to avoid sending too large chains of nodes as well as recursive
chains.