BackupShipping: force backups to do initial full sync
When an incremental backup is uploaded by a different node than any of the
previous backups came from, the drbd-metadata is no longer valid. Therefore
it must be recreated during a backup restore. By doing so, the usual
method for thin volumes can not be used, since we already have data but
lose the day0-GID by recreating the metadata.
To prevent not syncing the data to additional resources, any resource that
originated from a backup will now behave like a thick volume (by doing an
inital full sync) for all eternity - which also means making a regular
snapshot and restoring it will NOT get rid of this behavior.