BackupShipping: allow users to set forceInitialSync
A new property DrbdOptions/ForceInitialSync is now available. This property
forces resources to make a full sync even if they only have thin volumes.
While this behavior can be stopped by removing the property or setting it
to False, as soon as a resource has been created with this property active
it will be stuck with this behavior forever, even across snapshots.