Layer: prevent NPE when copying .meta
When vlm0 has external and vlm1 internal metadata, things used to break
because vlm0 was assumed to be the standard for every volume, leading to a
NullPointerException when trying to copy the non-existent external metadata
for vlm1.
Now, every volume checks whether it has external metadata and only tries
to copy it if it has.