BackupShipping: change process of choosing node for shipping
Finding all available nodes now groups the nodes according to their
metadata-layout. That means that a node whose resource has external
metadata will not be in the same group as a node whose resource has
internal metadata, and the same goes for various constellations involving
multiple volumes. This is necessary so that all nodes can ship the backup
without causing problems for an incremental shipping from a different node.

The other important change introduced is that now the list of available
nodes gets shuffled before a specific node is chosen (in cases where the
preferred node is not available). This allows for a more even distribution
of shipments across all possible nodes.