BackupShipping: l2l shipping rework
Previously l2l shipping worked like this:
1. The source cluster creates the snapshot and sets most of the
   props and flags required for the shipping (excluding mainly
   those needed to actually start the shipping), then creates the
   metadata and sends it to the target cluster
2. The target cluster creates the snapshot based on the metadata
   it received, prepares and starts the receive and checks
   its snapshots to find out whether an incremental shipping is
   possible (and on which snap it should be based)
3. The source cluster adds the last few missing flags & props to
   start the shipping, including props that specify if the shipping
   is incremental or not, then starts the shipping

Now, the process works like this:
a. The souce cluster finds all snapshots that could be used as a
   base for an incremental shipping and sends them to the target
   cluster
b. The target cluster decides if an incremental shipping can be
   done and which snapshot to use as a base for it
c. Using the information just received from the target cluster,
   the source cluster chooses a node to do the shipping and
   creates the new snapshot there, sets all flags and props
   except the ones that actually start the shipping and then
   creates and sends the metadata to the target cluster
d. The target cluster creates the snapshot based on the metadata
   it received and starts receiving
e. The source cluster adds the last few missing flags & props
   and starts the shipping

The big difference this change makes is it allows for a better
selection of a base snapshot. Since setting most of the props
needs the node that will do the shipping to be decided upon, in
(2) the target cluster had to have one of the snapshots present
on the source node, otherwise a full backup had to be made. Now,
however, the base snap is decided upon (a) before the node is
selected (c), and therefore restricts which node can be used as a
source node. This leads to far higher odds of an incremental backup
being possible, which reduced the shipping time.