Node,reconnect: Avoid "only one initialConnectSink allowed" error
When an API call wants to wait for a brief connection attempt to
the satellite, it uses the returned Flux to concat with additional
steps / info. There can only exist one such Flux.
If the connection could not be established that Flux was not
deleted (= set to null), which means if the initial connection
(after a `node create`) could not be established, the next
`node reconnect` will throw the mentioned ImplementationError.

This is now solved by deleting that initial Flux when the connection
could not be established after the timeout. Also instead of storing
only one FluxSink, multiple can be registered but should use
different keys since not all should be canceled if one fails.