Drbd: Change ready-condition
Until now 2 conditions had to be fulfilled for a resource to get "ready":

1) every local volume needed access to UpToDate data. That was either
possible if the volume is locally UpToDate (only possible when diskful)
or if the volume is connected to a peer that has this volume UpToDate.

2) the resource is connected to all known DRBD-peers.

The first condition makes sense and is still valid.
The second condition however could get in our way as the following scenario
shows:

3 Node setup, 2 diskful on nodeA and nodeB. nodeC does not have the
resource deployed.
nodeA and nodeB are both UpToDate.
nodeA goes offline (linstor + DRBD)
user tries to access nodeB's data from nodeC, i.e. creates a new
resource on nodeC. That 'resource create ..' command returns an
error with the message "Resource did not become ready in a reasonable
time", since nodeC could not connect to nodeA (condition 2) ).

Therefore this second condition was changed such that a newly created
resource only needs to connect to those peer's whos linstor node
is currently online. In other words, in the above example nodeC
only wants to connect to nodeB to mark itself as "ready".