DevMgr: actually delete nodes
A stlt remembers the remoteNodes that have replicas of
the local rscs it has deployed. When it removes the last
replica from one of those remoteNodes, it also deletes
the now empty remoteNode, since it does not need it for
anything else. This should also be the case when the
stlt deletes a local rsc that was the last remaining link
to a remoteNode.
However, that did not happen. Instead, it only removed
the remoteNode from the nodesMap, effectively forgetting
it knew that node. But this lead to a data leak, since
the node probably has a storPool, and that storPool has
volumeData, and that volumeData is linked to volumes and
snapshots and whatnot - and none of that is deleted, and
is also still accessible through at least the storPoolDfn.

A data leak is already bad enough, but now that the core
objects have equals and hashCode implemented, this problem
ended up leading to bigger consequences - namely an exception.
The scenario is somewhat specific. The following commands
produce an "attempt to replace an active transMgr" error:
  linstor r c bravo charlie delta rsc -s lvmpool
  linstor r d charlie rsc
  linstor r d bravo rsc
  linstor r c bravo charlie rsc -s lvmpool
  linstor r d charlie rsc
Bravo is the node the error will occur on. It's name is not
important. Charlie and delta, however, need to be in this
exact alphabetical order - the node you delete from (charlie)
needs to be sorted earlier than the node whose storPool
produces the error.
The way this leads to the error goes like this: for reasons
unknown charlie's rsc needs to be deleted, otherwise the error
does not trigger. Then we delete the rsc of the node we want the
error to happen on (bravo) which means bravo now has delta's
storPool (in bravo's storPoolDfn) without having delta in its nodesMap.
Then we create the two resources we deleted just now again. This
does the following: first, bravo creates node charlie and its
storPool again (since charlie comes first alphabetically). This
means the storPoolDfn needs to be added to the transMgr, which
also adds delta's old storPool to the transObjects-hashSet of the
TransactionObjectCollection.
After everything that bravo needs for charlie is done, its
delta's turn. So bravo creates delta's storPool and adds this new
storPool into the storPoolMap of the storPoolDfn, overwriting delta's
old storPool. It also tries to add delta's new storPool to the
transObjects-hashSet - but this does nothing, since both storPools
have the same hashCode (and equals).
So now we have two storpools, both have active transMgrs, but
only one of them gets committed in the end, since only the objects
in the transObjects-hashSet get committed. Which means, when we
try to delete charlies rsc, we will try to activate the new transMgr
for delta's new storPool (since we need to touch the storPoolDfn),
but this storPool still has an active transMgr - which leads to
the "attempting to replace an active transMgr" error.

The fix is simple enough - just delete the node instead of only
removing it from the nodesMap, this in turn deletes the storPool
and everything else that might have been left over previously.