DRBD: only delete .res files after full-sync
Prior this patch, the satellite deleted all .res files when it started.
In case of a network problem (or the controller not connecting to the
satellite), the satellite will never receive the resources required
to re-generate the .res files.

This behavior is now changed, such that the old .res files are only
deleted after the (linstor-) FullSync is applied and the first
run of the DeviceManager is about to start (which will regenerate
the .res files for the received resources).