Drbd,clone: Fix changed ready-condition
Commit d77d64b8aa731f69ab3f53053b1065cc76cdbc58 changed the ready-condition
for any kinds of resource creations (including spawn, autoplace and adjust)
but unfortunately not for cloning.

Cloning therefore still only checked / waited until each local DRBD peer
had local UpToDate data (diskless are not an issue since they are not cloned)
without waiting to connect to any other DRBD peer. If a test / plugin
immeadiately tried to run commands on the cloned DRBD resource, it simply
did so too fast, for example resulting in errors since DRBD was not connected.