Backups: fix normal snap restore
When restoring from a snapshot that was created by a backup create
(not by downloading the backup, just a normal snapshot restore),
this snapshot is also treated like a downloaded backup by recreating
the drbd-metadata and forcing an initial sync to happen.
However, for this to work linstor needs to forget that a primary
had already been decided upon, which previously was only done while
restoring from a downloaded backup, but is now done in both cases.