docker: use losetup-container shim
When using file based storage pools in containers, there is an issue
with the reported backing file of losetup. As a workaround, use
losetup-container as a shim for normal losetup.

The shim will try to find the correct backing file from inside the container,
making it "just work". Since the shim is only active for the specific
command line used by LINSTOR, it's important that the shim is kept
in sync with the invocation.