Rsc,delete: add delete hook
Any rsc that is added to the RetryResourcesTask needs to be removed
from it before it is deleted to prevent AccessToDeletedDataExceptions.
This also changes how the delete method of the rscDfn works - the
step of deleting all rscs now needs to happen outside of this
function.
It is strongly recommended that all future rsc-deletions happen only
through the newly added cleanupAndDelete method.