VlmDfn,discGran: Add automatic rs-discard-granularity calculation
The DeviceHandlerImpl traverses all AbsVlmData after processing. For every
AbsVlmData that has -1 (uninitialized) as discGran, lsblk is queried
for DISC-GRAN for the layer specific device of the current AbsVlmData. The
result is stored in AbsVlmData and sent to the controller, where the
rs-discard-granularity property can be calculated accordingly on the VlmDfn
level.

Additionally there is a separate property to enable or disable this feature
whose functionality is self-explanatory. Setting the rs-discard-granularity
property also triggers disabling the automatism of linstor managing that
property on the same level the rs-discard-granularity property was set.