Props: Add specialHandlers
There are more and more properties that require special handling
if they are set. This commit extends the fillProperties
and the remove methods of CtrlPropsHelper with a map of
(full-qualified) keys to Consumable that are executed if the
property is set. The consumable is called with a number of
parameters. One is the key and the value that triggered the
execution of the consumable as well as the old value. The new value
will be null if the property was deleted. The old value can be used to
perform a "changed" query that could determine if the consumable wants to
process the change or not.