Layer,size: Fixed calculating sizes
There were a few corner cases (i.e. with DRBD) where the data-device
needed to be rounded up to the next extent, but the meta-device
was just on the edge of the extent size (i.e. precisely 4MiB for LVM).
Since the data-device will be rounded up during the 'lvcreate ..'
command, the previously calculated size for the (external) meta-device
might be simply too small.

That means the previous size calculation allowed for the data-device
to be created larger as expected during the calculations leading
to a too small meta-device.

In order to fix this, all storage providers now cache their storage
pool local extent size to use it for a more accurate calculation.

In the very rare case such an extent size changes while linstor
is running (i.e. with 'vgchange -s ...') the satellite has to be
restarted in order to avoid the above mentioned corner cases.