Snap,abort: Fixed abort procedure
If an error occurs during the actual creation of a snapshot
(i.e. 'lvcreate --snapshot ...' ) the snapshot should be aborted and
for DRBD we also need to run resume-io. That is controlled by a
boolean in DrbdRscData, which is located on the resource.

Resuming IO is one step. Stopping a new attempt for creating a
snapshot is a different boolean that was not set until now
in the abort procedure.

Without disabling the "takeSnapshot" boolean, updating the
snapshot definition as well as updating the resource definition
afterwards also caused an error, preventing DRBD from actually
running the resume-io.

This patch fixes this issue by simply disabling "takeSnapshot".
Since the snapshot definition is updated first, no more attempts in
creating a snaptshot are made. The consequtive update of the
resource definition can also run the resume-io without being
prevented by a failed snapshot creation.