QSI: Add query-size-info API
This new API should replace the old query-max-volume-size since
the qmvs had to be kept compatible with the old autoplacer's
behavior (i.e. one storage pool name is selected for multiple
nodes instead of allowing different storage pool names for different
nodes).

Additionally this API returns an estimated available as well
as a capacity. These values are not perfectly accurate since they
are not based on the autoplacer itself and therefore are not taking
some limitations into consideration (like --replicas-on-same).