containers: fix semantic by using an if
Here I just had the wrong semantic which would hide a problem in the
"true" case, as in:

$ true && (true && false) || true

Which results in true, hiding the failure in the subshell while the
intention was to not fail overall if the first test was false.

What we actually want is something like this:

if true; then true && false; fi

Which results in false, but also handles the case when the test is
actually false.