add migration tools to linstor-config
Adds two new commands to linstor-config tools:

* run-migration applies any pending database migration, like the controller
  would on start up. Then it exits, without starting any additional services.
* all-migrations-applied checks whether there are any pending migrations.

Both these tools are useful for scripting.

A particular use case would be k8s: Before the controller actually starts,
we can have a quick check if any migrations need to be applied. The advantage
is that we can add additional backup steps before actually running the
migration. We also have a clear indication that the migration succeeded.