AutoSnap,BackupSchedule: Serialize multi-snapshots
The issue with concurent many snapshots is that the satellites are
optimized to parallelize processing, which means that multiple snapshots
in the same DevMgr run will result in first all snapshots being
suspend-io'd, all snapshots are created and afterwards all snapshots are
resume-io'd. That means that the time the resources are in suspend-io state
is scaling with the number of snapshots. Since taking a snapshot takes
roughly about a second, 100+ snapshots can easily suspend IO for the
resources for a minute or more which is usually not taken well by the
applications running above DRBD.

This patch extends the existing BackgroundRunner such that the given Flux
is only executed if it can take all required node-locks. If no node-lock
is specified (default), the BackgroundRunner behaves the same as before.

After the Flux is finished, a method is called to check if a queued Flux
can be started. Doing so, all RunConfig objects have both, a priority as
well as an ID. The ID is only used as a tiebreaker when the priorities
are even.

Some examples:
Flux 1 requires nodes A, B and C. Since this is the first Flux, it can
start immediately.
Flux 2 requires nodes C and D. Since node C is still occupied, Flux 2 has
to wait.
Flux 3 requires nodes E and F. No conflict here, so Flux 3 can start.
Flux 4 requires only node D. Even if node D is currently not occupied, if
Flux 2 has a higher priority (which it has by default), Flux 4 is still
queued since Flux 2 has reserved node D already.