MultiSnap: Add support for multiple snapshots
The new API allows to specify a list of snapshots (each with
different nodes, resource name and snapshot name). These snapshots
are handled atomically which means that the controller needs to
make sure that all these requests / updates are always processed
in the same DevMgrRun on the satellite.

This is achieved with a new internal API ChangedData, which currently
only supports resources and snapshots, but can be easily extended for
nodes, storage pools, etc...