QSI: Specify required fields
The QuerySizeInfoRequest does not require an autoSelectFilter.
In that case the selectFilter from the RG is used.

The QuerySizeInfoResponse usually returns a space_info, but not
guaranteed (i.e. if no storage pools are available at all). In
that case corresponding reports are to be expected. If all goes
well, reports are not expected but only the space_info.
Once we update the generator to add
@JsonInclude(JsonInclude.Include.ALWAYS) annotations, we might
want to consider making 'reports' required.

If the QuerySizeInfoResponseSpaceInfo exists, max_vlm_size_in_kib
is required (might however be 0). 'next_spawn_result' could also
become required as soon as we improve the generator to include
(required) empty arrays in the response.
The other fields might become optional in the future to skip
unnecessary simulations.

Both fields of the QuerySizeInfoSpawnResults are requried since
they describe a storage pool.