BackupShipping: make queues work with l2l-shipping
To decide whether an item from the queue can be started or
needs to be added back to the queue, it needs to be confirmed if
its prevSnap still exists. For s3, this is easy enough, but with
l2l this involves asking the target-cluster for information on said
prevSnap. This leads to the whole process becoming asynchronous in
case of an l2l-shipping - which means normal loops won't be able to
decide whether or not an additional shipping needs to be started
since the request to the target cluster has not been answered yet.

To solve this problem (and make sure that s3 still contiues as fast
as before) a flux-loop is needed. Basically, as soon as the
target-cluster answers it is not only decided what to do with the
current queue-item, but also whether or not to try to start the next
item (which depends on the number of free shipping slots, which could
have changed depending on whether or not the current item can be
started). While a bit convoluted, this logic ensures that no node goes
above its allowed concurrent shippings count.