ETCD,Layer: Fix broken key calculation
As the new method "buildPrefixedRscLayerIdKey" in BaseEtcdDriver
explains, loading all volumes for a given resource-layer-id was
done as follows:
1) build a key that is roughly "/LINSTOR/<table>/<rscId>/"
2) cut the last "/"
3) load all entries with the resulting string as "prefix" (i.e.
load all entries starting with the given key)

The problem was, that this method also loaded volumes of other
rscIds. For example if our rscId is 2, a ".../2" prefix will
also match ".../20", ".../22", etc.

This patch fixes this issue with a new step between 2 and 3
that adds a primary-key delimiter ":" to the end of the key.