MultiSnap: get suspend-ios closer together
Before this commit, the DeviceHandlerImpl iterated the resources
one after the other, completely processing the resource. Usually
a required suspend-io would be one of the first things to check,
but the rest of the resource (including all of its layers) would
follow before the next resource has the chance to suspend-io itself.
This commits adds an additional loop before the main processing
that only suspends / or resumes IO on the topmost layer if needed.
After the loop, the unchanged processing loop runs as usual.