BackupShipping: limit luks to one node
Due to the fact that every volume is encrypted with a different
volume key, backup shippings with luks in the layer tree can not
be continued by a different node than the one who made the full
backup. Therefore, backup shippings with luks now behave the same
as with zfs.