BackupShipping: only abort the snapDfn, not the full node
When aborting a backup shipping, the logic used to get the node on which
the shipping was and abort all shippings on that node. This meant that any
shippings on that node would be aborted, whether they should have been or
not.
This patch fixes the issue by only aborting the given snapDfn.